data-collection module/requirements.txt
# Data Collection Module Requirements

# Core data fetching
yfinance>=0.2.30
pandas>=2.0.0
numpy>=1.24.0

# Optional: Backup data sources
twelvedata>=1.2.8  # Twelve Data API (optional)
nsepy>=0.8  # NSE India data (optional, for daily fallback)

# Data storage
pyarrow>=14.0.0  # For Parquet format
fastparquet>=2023.10.0

# Scheduling
APScheduler>=3.10.0

# Environment variables
python-dotenv>=1.0.0

# Logging and utilities
tqdm>=4.66.0

# Optional: Parallel processing
joblib>=1.3.0
